Updating doco for auto subscribe
Porting autosubscribe doco to v4

Porting pubsub cleanup to v3

Fixes after rebase

Adding .DoNotrequireRouting to the upgrade guide

Added .orig to gitignore

Fix after rebase

More fixes after rebase

Remove the cruft added by the host package

Removed extra host from ref

Fixes

More cleanup

Kill new lines
