--- Snippets/Snippets_4/SubscriptionSettings.cs
@@ -1,18 +1,0 @@
-﻿namespace Snippets4
-{
-    using NServiceBus;
-    using NServiceBus.Features;
-
-    public class SubscriptionSettings
-    {
-        public void DisableAutoSubscribe()
-        {
-            #region DisableAutoSubscribe
-
-            Configure.Features.Disable<AutoSubscribe>();
-
-            #endregion
-        }
-
-    }
-}
