+++ Snippets/Snippets_5/Routing/BasicOperations.cs
@@ -1,4 +1,4 @@
-﻿namespace Snippets6.Routing
+﻿namespace Snippets5.Routing
 {
     using NServiceBus;
