+++ Snippets/Snippets_5/Routing/EndpointMapping/ConfigurationSource/ConfigurationSource.cs
@@ -1,11 +1,10 @@
-﻿namespace Snippets4.EndpointMapping.ConfigurationSource
+﻿namespace Snippets5.Routing.EndpointMapping.ConfigurationSource
 {
     using System.Configuration;
     using NServiceBus.Config;
     using NServiceBus.Config.ConfigurationSource;
 
     #region endpoint-mapping-configurationsource
-
     public class ConfigurationSource : IConfigurationSource
     {
         public T GetConfiguration<T>() where T : class, new()
@@ -13,7 +12,7 @@
             if (typeof(T) == typeof(UnicastBusConfig))
             {
                 //read from existing config 
-                var config = (UnicastBusConfig) ConfigurationManager
+                var config = (UnicastBusConfig)ConfigurationManager
                     .GetSection(typeof(UnicastBusConfig).Name);
                 if (config == null)
                 {
@@ -37,6 +36,5 @@
             return ConfigurationManager.GetSection(typeof(T).Name) as T;
         }
     }
-
     #endregion
 }
