--- Snippets/Snippets_5/SubscriptionSettings.cs
@@ -1,20 +1,0 @@
-﻿namespace Snippets5
-{
-    using NServiceBus;
-    using NServiceBus.Features;
-
-    public class SubscriptionSettings
-    {
-        public void DisableAutoSubscribe()
-        {
-            #region DisableAutoSubscribe
-
-            BusConfiguration busConfiguration = new BusConfiguration();
-
-            busConfiguration.DisableFeature<AutoSubscribe>();
-
-            #endregion
-        }
-
-    }
-}
