+++ Snippets/Snippets_6/Routing/EndpointMapping/ConfigurationSource/ConfigurationSource.cs
@@ -32,7 +32,6 @@
                 return config as T;
             }
 
-            // To in app.config for other sections not defined in this method, otherwise return null.
             return ConfigurationManager.GetSection(typeof(T).Name) as T;
         }
     }
