--- Snippets/Snippets_6/Routing/EndpointMapping/ConfigureEndpointMapping.config
@@ -1,14 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<!--startcode endpoint-mapping-appconfig-->
-<configuration>
-  <configSections>
-    <section name="UnicastBusConfig"
-        type="NServiceBus.Config.UnicastBusConfig, NServiceBus.Core"/>
-  </configSections>
-  <UnicastBusConfig >
-    <MessageEndpointMappings>
-      <add Assembly="assembly" Endpoint="queue@machinename" />
-    </MessageEndpointMappings>
-  </UnicastBusConfig>
-</configuration>
-<!--endcode-->
