+++ Snippets/Snippets_6/Scanning/ScanningPublicApi.cs
@@ -45,8 +45,5 @@
         }
 
         #endregion
-
-        interface IConfigureThisEndpoint
-        {}
     }
 }
