+++ Snippets/Snippets_6/packages.config
@@ -1,8 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NServiceBus" version="6.0.0-unstable0601" targetFramework="net451" />
+  <package id="NServiceBus" version="6.0.0-unstable0612" targetFramework="net451" />
   <package id="NServiceBus.Callbacks" version="1.0.0-unstable0011" targetFramework="net451" />
   <package id="NServiceBus.Host" version="7.0.0-pullrequest0040" targetFramework="net451" />
-  <package id="NServiceBus.Host" version="6.0.0" targetFramework="net45" />
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
 </packages>
