+++ nservicebus/upgrades/5to6.md
@@ -164,3 +164,7 @@ When you wanted to use the `IBuilder` interface outside the infrastructure of NS
 `var builder = ((UnicastBus)bus).Builder`
 
 This is no longer supported. We advise to, instead of using `IBuilder` directly, use dependency injection via your container of choice.
+
+## AutoSubscribe
+
+The configuration option `DoNotRequireExplicitRouting()` has been obsoleted since transports with support for centralized pubsub will always auto subscribe all events without requiring explicit routing. Transports with message driven pubsub (like MSMQ, SqlServer and AzureStorageQueues) will not subscribe properly if there is no routing specified. If you used this option you can safely be removed it.
